refactor(catalog): delete the rollup pipeline
The 33-into-11 mapper, the queries that served it, and the write path that
regenerated tags from it are gone. Ports first, then their implementations —
reflection DI hides an impl-first mistake until runtime.

Normalize moves to the key parser on its way past: folding diacritics to match
a piucenter key against a chart never had anything to do with skills, and it
was only living on the mapper because that is where it was written.

The crawl now ends at metrics and baselines. Nothing derives an abstract tag
from a measurement any more.
