--- ScoreTracker/ScoreTracker.Catalog/Contracts/Queries/GetChartSkillsQuery.cs
@@ -1,10 +1,0 @@
-﻿using MediatR;
-using ScoreTracker.Domain.Records;
-
-namespace ScoreTracker.Catalog.Contracts.Queries
-{
-    [ExcludeFromCodeCoverage]
-    public sealed record GetChartSkillsQuery : IQuery<IEnumerable<ChartSkillsRecord>>
-    {
-    }
-}
