--- ScoreTracker/ScoreTracker.Domain/Records/ChartSkillsRecord.cs
@@ -1,10 +1,0 @@
-﻿using ScoreTracker.SharedKernel.Enums;
-
-namespace ScoreTracker.Domain.Records
-{
-    [ExcludeFromCodeCoverage]
-    public sealed record ChartSkillsRecord(Guid ChartId, IEnumerable<Skill> ContainsSkills,
-        IEnumerable<Skill> HighlightsSkill)
-    {
-    }
-}
